+++ packages/config/src/bindings.ts
@@ -1,46 +1,12 @@
-// oxlint-disable typescript/no-explicit-any -- needed in type utilities
-
+import type {
+	ConfigOrUndefined,
+	DurableObjectExportName,
+	UnwrapConfig,
+	WorkerExportName,
+	WorkerName,
+} from "./inference";
 import type { Json } from "./utils";
-import type { Pipeline, PipelineRecord } from "cloudflare:pipelines";
-
-// ═══════════════════════════════════════════════════════════════════════════
-// GENERIC UTILITIES
-// ═══════════════════════════════════════════════════════════════════════════
-
-/**
- * The Worker's entry module, imported with the `{ type: "cf-worker" }` import attribute
- * @example
- * ```ts
- * import * as entrypoint from "./src" with { type: "cf-worker" };
- * ```
- */
-export type WorkerModule = Record<string, any>;
-
-/**
- * Default module type representing an unknown Worker's exports.
- * - default export can be `ExportedHandler` or a `WorkerEntrypoint` class constructor
- * - named exports can be `WorkerEntrypoint`, `DurableObject`, or `WorkflowEntrypoint` class constructors
- */
-interface DefaultModule {
-	default?: ExportedHandler | Constructor<Rpc.WorkerEntrypointBranded>;
-	[key: string]:
-		| ExportedHandler
-		| Constructor<Rpc.WorkerEntrypointBranded>
-		| Constructor<Rpc.DurableObjectBranded>
-		| Constructor<Rpc.WorkflowEntrypointBranded>
-		| undefined;
-}
-
-/**
- * Represents a class constructor that creates instances of TInstance.
- */
-type Constructor<TInstance> = new (...args: any[]) => TInstance;
-
-/**
- * Extracts the instance type from a class constructor if it extends `TInstance`.
- */
-type ExtractInstance<T, TInstance> =
-	T extends Constructor<TInstance> ? InstanceType<T> : never;
+import type { PipelineRecord } from "cloudflare:pipelines";
 
 // ═══════════════════════════════════════════════════════════════════════════
 // BINDING OPTIONS
@@ -597,341 +563,6 @@ export interface WorkflowBinding<
 }
 
 // ═══════════════════════════════════════════════════════════════════════════
-// BINDING TYPE INFERENCE
-// Maps binding definitions to their corresponding Cloudflare runtime types.
-// ═══════════════════════════════════════════════════════════════════════════
-
-/**
- * Mapping from binding type literals to Cloudflare runtime types.
- *
- * For parameterized bindings (ai, kv, queue etc.), the type parameter is extracted
- * from the binding's `__typeParams` property.
- */
-interface BindingTypeMap<TBinding> {
-	// Parameterized bindings - extract `__typeParams` and apply to runtime types
-	ai: TBinding extends { __typeParams: [infer T extends AiModelListType] }
-		? Ai<T>
-		: Ai;
-	kv: TBinding extends { __typeParams: [infer T extends string] }
-		? KVNamespace<T>
-		: KVNamespace;
-	pipeline: TBinding extends { __typeParams: [infer T extends PipelineRecord] }
-		? Pipeline<T>
-		: Pipeline;
-	queue: TBinding extends { __typeParams: [infer T] } ? Queue<T> : Queue;
-
-	// Non-parameterized bindings
-	"ai-search": AiSearchInstance;
-	"ai-search-namespace": AiSearchNamespace;
-	"analytics-engine-dataset": AnalyticsEngineDataset;
-	artifacts: Artifacts;
-	assets: Fetcher;
-	browser: Fetcher;
-	d1: D1Database;
-	"dispatch-namespace": DispatchNamespace;
-	flagship: Flagship;
-	hyperdrive: Hyperdrive;
-	images: ImagesBinding;
-	logfwdr: any;
-	media: MediaBinding;
-	"mtls-certificate": Fetcher;
-	"rate-limit": RateLimit;
-	r2: R2Bucket;
-	secret: string;
-	"secrets-store-secret": SecretsStoreSecret;
-	"send-email": SendEmail;
-	stream: StreamBinding;
-	vectorize: VectorizeIndex;
-	"version-metadata": WorkerVersionMetadata;
-	"vpc-service": Fetcher;
-	"vpc-network": Fetcher;
-	unsafe: any;
-	"worker-loader": WorkerLoader;
-}
-
-type InferBindingType<TBinding> =
-	// Worker binding
-	TBinding extends {
-		type: "worker";
-		exportName: infer TExportName extends string;
-		__config: infer TUnwrappedConfig;
-	}
-		? InferMainModule<TUnwrappedConfig> extends infer TModule extends
-				WorkerModule
-			? TExportName extends keyof TModule
-				? TModule[TExportName] extends Constructor<any>
-					? Fetcher<
-							ExtractInstance<TModule[TExportName], Rpc.WorkerEntrypointBranded>
-						>
-					: Fetcher
-				: never
-			: never
-		: // Durable Object binding
-			TBinding extends {
-					type: "durable-object";
-					exportName: infer TExportName extends string;
-					__config: infer TUnwrappedConfig;
-			  }
-			? InferMainModule<TUnwrappedConfig> extends infer TModule extends
-					WorkerModule
-				? TExportName extends keyof TModule
-					? DurableObjectNamespace<
-							ExtractInstance<TModule[TExportName], Rpc.DurableObjectBranded>
-						>
-					: never
-				: never
-			: // Workflow binding
-				TBinding extends {
-						type: "workflow";
-						exportName: infer TExportName extends string;
-						__config: infer TUnwrappedConfig;
-				  }
-				? InferMainModule<TUnwrappedConfig> extends infer TModule extends
-						WorkerModule
-					? TExportName extends keyof TModule
-						? ExtractInstance<
-								TModule[TExportName],
-								Rpc.WorkflowEntrypointBranded
-							> extends infer TWorkflow
-							? TWorkflow extends {
-									run(event: { payload: infer P }, step: any): any;
-								}
-								? Workflow<P>
-								: Workflow
-							: Workflow
-						: never
-					: never
-				: // JSON bindings
-					TBinding extends { type: "json"; value: infer TValue }
-					? TValue
-					: // Text bindings
-						TBinding extends { type: "text"; value: infer TValue }
-						? TValue
-						: // Other bindings
-							TBinding extends {
-									type: infer K extends keyof BindingTypeMap<TBinding>;
-							  }
-							? BindingTypeMap<TBinding>[K]
-							: never;
-
-// ═══════════════════════════════════════════════════════════════════════════
-// CROSS-WORKER BINDING HELPERS (INTERNAL)
-// Types used by the Bindings interface for type-safe cross-worker bindings.
-// ═══════════════════════════════════════════════════════════════════════════
-
-/**
- * Infer the Worker name from a config.
- *
- * @example
- * ```typescript
- * import { defineConfig } from "@cloudflare/config";
- * import type { InferDurableNamespaces, UnwrapConfig } from "@cloudflare/config";
- *
- * const config = defineConfig({ name: "my-worker", ... });
- *
- * type WorkerConfig = UnwrapConfig<typeof config>;
- * // Inferred as: "my-worker"
- * type Name = InferWorkerName<WorkerConfig>;
- * ```
- */
-type InferWorkerName<TUnwrappedConfig> = TUnwrappedConfig extends {
-	name: infer TName extends string;
-}
-	? TName
-	: never;
-
-/**
- * Extract config(s) from `TUnwrappedConfig` where `TName` is assignable to the config's name.
- * Unlike Extract, this works when a config has a union name like `"worker-a"` | `"worker-b"`.
- */
-type ExtractConfigByName<TUnwrappedConfig, TName extends string> =
-	TName extends InferWorkerName<TUnwrappedConfig> ? TUnwrappedConfig : never;
-
-/**
- * Resolves the config type for cross-worker bindings.
- * Returns the extracted config if `workerName` matches a known config.
- * Otherwise returns `undefined`.
- */
-type ConfigOrUndefined<TUnwrappedConfig, TName extends string> =
-	TName extends InferWorkerName<TUnwrappedConfig>
-		? ExtractConfigByName<TUnwrappedConfig, TName>
-		: undefined;
-
-/**
- * Returns valid Worker names.
- * For typed configs, returns the constrained union of known Worker names.
- * Otherwise, returns `string` to allow any `Worker` name.
- */
-type WorkerName<TUnwrappedConfig> = TUnwrappedConfig extends { name: string }
-	? InferWorkerName<TUnwrappedConfig>
-	: string;
-
-/**
- * Infer export names from a config's exports, optionally filtered by type.
- * When TExportType is `string` (default), returns all export names.
- * When TExportType is a specific literal like `"durable-object"` or `"workflow"`,
- * returns only exports of that type.
- */
-type InferExportsByType<
-	TUnwrappedConfig,
-	TExportType extends string = string,
-> = TUnwrappedConfig extends {
-	exports: infer TExports extends Record<string, { type: string }>;
-}
-	? {
-			[K in keyof TExports]: TExports[K] extends { type: TExportType }
-				? K & string
-				: never;
-		}[keyof TExports]
-	: never;
-
-/**
- * Infer `WorkerEntrypoint` export names from a config.
- * Returns named module exports that are not declared as type `"durable-object"` or `"workflow"` in `exports`.
- * Excludes `"default"` since `exportName` should only be provided for named exports.
- */
-type InferEntrypointExports<TUnwrappedConfig> = Exclude<
-	keyof InferMainModule<TUnwrappedConfig> & string,
-	| "default"
-	| InferExportsByType<TUnwrappedConfig, "durable-object" | "workflow">
->;
-
-/**
- * Returns valid exports for a Worker binding.
- * For known Workers, returns the constrained union of entrypoint exports.
- * For unknown Workers in typed configs, returns `never`.
- * For untyped configs, returns `string` to allow any export name.
- */
-type WorkerExportName<
-	TUnwrappedConfig,
-	TName extends string,
-> = TUnwrappedConfig extends { name: string }
-	? TName extends InferWorkerName<TUnwrappedConfig>
-		? InferEntrypointExports<ExtractConfigByName<TUnwrappedConfig, TName>>
-		: never
-	: string;
-
-/**
- * Returns valid exports for a Durable Object binding.
- * For known Workers, returns the constrained union of Durable Object exports.
- * For unknown Workers in typed configs, returns `never`.
- * For untyped configs, returns `string` to allow any export name.
- */
-type DurableObjectExportName<
-	TUnwrappedConfig,
-	TName extends string,
-> = TUnwrappedConfig extends { name: string }
-	? TName extends InferWorkerName<TUnwrappedConfig>
-		? InferExportsByType<
-				ExtractConfigByName<TUnwrappedConfig, TName>,
-				"durable-object"
-			>
-		: never
-	: string;
-
-// TODO: re-enable when workflow bindings return.
-// /**
-//  * Returns valid exports for a Workflow binding.
-//  * For known Workers, returns the constrained union of Workflow exports.
-//  * For unknown workers in typed configs, returns `never`.
-//  * For untyped configs, returns `string` to allow any export name.
-//  */
-// type WorkflowExportName<
-// 	TUnwrappedConfig,
-// 	TName extends string,
-// > = TUnwrappedConfig extends { name: string }
-// 	? TName extends InferWorkerName<TUnwrappedConfig>
-// 		? InferExportsByType<
-// 				ExtractConfigByName<TUnwrappedConfig, TName>,
-// 				"workflow"
-// 			>
-// 		: never
-// 	: string;
-
-// ═══════════════════════════════════════════════════════════════════════════
-// CONFIG INFERENCE (PUBLIC)
-// Exported types for inferring runtime types from config definitions.
-// ═══════════════════════════════════════════════════════════════════════════
-
-/**
- * Unwrap function and promise types to get the underlying config.
- * Use this to normalize a config before passing it to other inference utilities.
- */
-export type UnwrapConfig<TConfig> = TConfig extends (
-	...args: any[]
-) => infer TReturn
-	? UnwrapConfig<TReturn>
-	: TConfig extends Promise<infer TCompletion>
-		? TCompletion
-		: TConfig;
-
-/**
- * Infer the `Env` interface type from a Worker config.
- *
- * Transforms a config object's `env` bindings into their
- * corresponding Cloudflare runtime types.
- *
- * @example
- * ```typescript
- * import { defineConfig, bindings } from "@cloudflare/config";
- * import type { InferEnv, UnwrapConfig } from "@cloudflare/config";
- *
- * const config = defineConfig({
- *   env: {
- *     MY_JSON: bindings.json({ id: string }),
- *     MY_KV: bindings.kv(),
- *   },
- * });
- *
- * type WorkerConfig = UnwrapConfig<typeof config>;
- * // Inferred as: { MY_JSON: { id: string }; MY_KV: KVNamespace }
- * export type Env = InferEnv<WorkerConfig>;
- * ```
- */
-export type InferEnv<TUnwrappedConfig> = TUnwrappedConfig extends {
-	env: infer TEnv extends Record<string, any>;
-}
-	? { [K in keyof TEnv]: InferBindingType<TEnv[K]> }
-	: never;
-
-/**
- * Infer the Durable Object namespace names from a Worker config's exports.
- * Returns a union of export names that have `type: "durable-object"`.
- *
- * @example
- * ```typescript
- * import { defineConfig } from "@cloudflare/config";
- * import type { InferDurableNamespaces, UnwrapConfig } from "@cloudflare/config";
- *
- * const config = defineConfig({
- *   exports: {
- *     MyDurableObject: { type: "durable-object", storage: "sqlite" },
- *     MyWorkflow: { type: "workflow", name: "my-workflow" },
- *   },
- * });
- *
- * type WorkerConfig = UnwrapConfig<typeof config>;
- * // Inferred as: "MyDurableObject"
- * type DurableNamespaces = InferDurableNamespaces<WorkerConfig>;
- * ```
- */
-export type InferDurableNamespaces<TUnwrappedConfig> = InferExportsByType<
-	TUnwrappedConfig,
-	"durable-object"
->;
-
-/**
- * Infer the main module type from a Worker config's entrypoint.
- * If entrypoint is a module namespace object, returns that type.
- * If entrypoint is a `string` or not present, returns `DefaultModule` as a fallback.
- */
-export type InferMainModule<TUnwrappedConfig> = TUnwrappedConfig extends {
-	entrypoint: infer TModule extends WorkerModule;
-}
-	? TModule
-	: DefaultModule;
-
-// ═══════════════════════════════════════════════════════════════════════════
 // BINDINGS API
 // The public interface and factory for creating binding definitions.
 // ═══════════════════════════════════════════════════════════════════════════
@@ -1301,229 +932,3 @@ export function createBindings<TConfig>(): Bindings<TConfig> {
  * ```
  */
 export const bindings = createBindings();
-
-// ═══════════════════════════════════════════════════════════════════════════
-// TRIGGERS API
-// Named types and helper factories for declaring event triggers.
-// ═══════════════════════════════════════════════════════════════════════════
-
-interface FetchTriggerOptions {
-	/**
-	 * A route that your Worker should be published to.
-	 *
-	 * For reference, see https://developers.cloudflare.com/workers/wrangler/configuration/#types-of-routes
-	 */
-	pattern: string;
-	/**
-	 * The DNS zone the pattern is attached to. Required when the
-	 * pattern is ambiguous.
-	 */
-	zone?: string;
-}
-
-interface QueueConsumerTriggerOptions {
-	/** The name of the queue from which this consumer should consume. */
-	name: string;
-	/** The queue to send messages that failed to be consumed. */
-	deadLetterQueue?: string;
-	/** The maximum number of messages per batch. */
-	maxBatchSize?: number;
-	/** The maximum number of seconds to wait to fill a batch with messages. */
-	maxBatchTimeout?: number;
-	/**
-	 * The maximum number of concurrent consumer Worker invocations.
-	 * Leaving this unset will allow your consumer to scale to the
-	 * maximum concurrency needed to keep up with the message backlog.
-	 */
-	maxConcurrency?: number | null;
-	/** The maximum number of retries for each message. */
-	maxRetries?: number;
-	/** The number of seconds to wait before retrying a message. */
-	retryDelay?: number;
-	/** The number of milliseconds to wait for pulled messages to become visible again. */
-	visibilityTimeoutMs?: number;
-}
-
-interface ScheduledTriggerOptions {
-	/**
-	 * "Cron" definitions to trigger a Worker's "scheduled" function.
-	 *
-	 * Lets you call Workers periodically, much like a cron job.
-	 *
-	 * More details here https://developers.cloudflare.com/workers/platform/cron-triggers
-	 */
-	schedules: string[];
-}
-
-/**
- * Fetch trigger — a route that your Worker should be published to.
- *
- * For reference, see https://developers.cloudflare.com/workers/wrangler/configuration/#types-of-routes
- */
-export interface FetchTrigger extends FetchTriggerOptions {
-	type: "fetch";
-}
-
-/**
- * Queue consumer trigger — invokes this Worker when messages arrive on the
- * named queue.
- *
- * For reference, see https://developers.cloudflare.com/workers/wrangler/configuration/#queues
- */
-export interface QueueConsumerTrigger extends QueueConsumerTriggerOptions {
-	type: "queue";
-}
-
-/**
- * Scheduled (cron) trigger — invokes this Worker on the given schedules.
- *
- * More details here https://developers.cloudflare.com/workers/platform/cron-triggers
- */
-export interface ScheduledTrigger extends ScheduledTriggerOptions {
-	type: "scheduled";
-}
-
-/**
- * Event triggers — fetch routes, queue consumers, and cron schedules
- * — that invoke this Worker. Construct entries with `triggers.fetch(...)`,
- * `triggers.queue(...)`, or `triggers.scheduled(...)`.
- *
- * For reference, see https://developers.cloudflare.com/workers/wrangler/configuration/#triggers
- */
-export interface Triggers {
-	/**
-	 * Fetch trigger — a route that your Worker should be published to.
-	 *
-	 * For reference, see https://developers.cloudflare.com/workers/wrangler/configuration/#types-of-routes
-	 */
-	fetch(options: FetchTriggerOptions): FetchTrigger;
-	/**
-	 * Queue consumer trigger — invokes this Worker when messages arrive on the
-	 * named queue.
-	 *
-	 * For reference, see https://developers.cloudflare.com/workers/wrangler/configuration/#queues
-	 */
-	queue(options: QueueConsumerTriggerOptions): QueueConsumerTrigger;
-	/**
-	 * Scheduled (cron) trigger — invokes this Worker on the given schedules.
-	 *
-	 * More details here https://developers.cloudflare.com/workers/platform/cron-triggers
-	 */
-	scheduled(options: ScheduledTriggerOptions): ScheduledTrigger;
-}
-
-/**
- * Triggers builder for configuring event triggers.
- *
- * @example
- * ```typescript
- * import { defineConfig, triggers } from "@cloudflare/config";
- *
- * export default defineConfig({
- *   triggers: [
- *     triggers.fetch({ pattern: "example.com/*", zone: "example.com" }),
- *     triggers.queue({ name: "my-queue" }),
- *     triggers.scheduled({ schedules: ["0 * * * *"] }),
- *   ],
- * });
- * ```
- */
-export const triggers: Triggers = {
-	fetch: (options) => ({ type: "fetch", ...options }),
-	queue: (options) => ({ type: "queue", ...options }),
-	scheduled: (options) => ({ type: "scheduled", ...options }),
-};
-
-// ═══════════════════════════════════════════════════════════════════════════
-// EXPORTS API
-// Named types and helper factories for declaring class exports.
-// ═══════════════════════════════════════════════════════════════════════════
-
-interface DurableObjectExportOptions {
-	/**
-	 * Storage backend for the Durable Object.
-	 *
-	 * - `"sqlite"`: selects the SQLite-backed storage engine
-	 *   (recommended for new classes).
-	 * - `"legacy-kv"`: selects the legacy key-value storage engine.
-	 */
-	storage: "sqlite" | "legacy-kv";
-}
-
-interface WorkflowExportOptions {
-	/** The name of the Workflow. */
-	name: string;
-	/** Optional limits for the Workflow. */
-	limits?: {
-		/** Maximum number of steps a Workflow instance can execute. */
-		steps?: number;
-	};
-}
-
-/**
- * Declares a Durable Object class defined by this Worker.
- *
- * For more information about Durable Objects, see the documentation at
- * https://developers.cloudflare.com/workers/learning/using-durable-objects
- *
- * For reference, see https://developers.cloudflare.com/workers/wrangler/configuration/#durable-objects
- */
-export interface DurableObjectExport extends DurableObjectExportOptions {
-	type: "durable-object";
-}
-
-/** Declares a Workflow defined by this Worker. */
-export interface WorkflowExport extends WorkflowExportOptions {
-	type: "workflow";
-}
-
-/**
- * Configuration for named exports declared by the Worker. Each entry's
- * key is the exported class name; the value configures the export.
- * Construct entries with `exports.durableObject(...)` or
- * `exports.workflow(...)`.
- *
- * Two export kinds are supported:
- *
- * - `durable-object`: declares a Durable Object class defined by this
- *   Worker. For more information about Durable Objects, see the
- *   documentation at
- *   https://developers.cloudflare.com/workers/learning/using-durable-objects
- *
- *   For reference, see https://developers.cloudflare.com/workers/wrangler/configuration/#durable-objects
- *
- * - `workflow`: declares a Workflow defined by this Worker.
- */
-export interface Exports {
-	/**
-	 * Declares a Durable Object class defined by this Worker.
-	 *
-	 * For more information about Durable Objects, see the documentation at
-	 * https://developers.cloudflare.com/workers/learning/using-durable-objects
-	 *
-	 * For reference, see https://developers.cloudflare.com/workers/wrangler/configuration/#durable-objects
-	 */
-	durableObject(options: DurableObjectExportOptions): DurableObjectExport;
-	/** Declares a Workflow defined by this Worker. */
-	workflow(options: WorkflowExportOptions): WorkflowExport;
-}
-
-/**
- * Exports builder for configuring Worker exports.
- *
- * @example
- * ```typescript
- * import { defineConfig, exports } from "@cloudflare/config";
- *
- * export default defineConfig({
- *   exports: {
- *     MyDurableObject: exports.durableObject({ storage: "sqlite" }),
- *     MyWorkflow: exports.workflow({ name: "my-workflow" }),
- *   },
- * });
- * ```
- */
-export const exports: Exports = {
-	durableObject: (options) => ({ type: "durable-object", ...options }),
-	workflow: (options) => ({ type: "workflow", ...options }),
-};
